Chore: #10 유저 관리와 카카오 API 설정 파일 분리
